Backend/BeetleUserAuth/requirements.txt
asgiref==3.4.1
astroid==2.5.6
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
Django==3.2.5
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
dnspython==2.2.0
idna==3.3
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mongoengine==0.23.1
mypy==0.931
mypy-extensions==0.4.3
pip==21.2.4
psycopg2==2.8.6
psycopg2-binary==2.9.3
PyJWT==2.3.0
pylint==2.8.2
pylint-django==2.5.2
pylint-plugin-utils==0.7
pymongo==3.12.3
pytz==2021.3
requests==2.27.1
setuptools==58.0.4
sqlparse==0.2.4
toml==0.10.2
tomli==2.0.1
typing-extensions==3.10.0.2
urllib3==1.26.8
wheel==0.37.1
wincertstore==0.2
wrapt==1.12.1
